Updated as we want all the results, all the time
Updated as we want all the results, all the time.   GetAndValidateLocationTimeWindow could also be modified since "returnAllResults" should always be true and this could be removed from GetAndValidateLocationTimeWindow.  However, committing this to the PR so others can test it against WattTime scenarios to see if there is something that breaks it here.